+++ R/AFM.lineProfile.R
@@ -5,10 +5,11 @@
 #'   points are provided in units of nanometers
 #'
 #' @param obj AFMdata object
-#' @param x1 start x position in  units of nm from bottom left
-#' @param y1 start y position in  units of nm from bottom left
-#' @param x2 end x position in  units of nm from bottom left
-#' @param y2 end y position in  units of nm from bottom left
+#' @param x1 start x position in  units of nm/pixels from bottom left
+#' @param y1 start y position in  units of nm/pixels from bottom left
+#' @param x2 end x position in  units of nm/pixels from bottom left
+#' @param y2 end y position in  units of nm/pixels from bottom left
+#' @param unitPixels logical, if TRUE, then coordinates are in units of pixels otherwise nm
 #' @param verbose logical, if \code{TRUE}, output additional information
 #' @returns AFMdata object with line data
 #'
@@ -18,27 +19,62 @@
 #' afmd = AFM.artificialImage(width=128, height=128, type='calibration', verbose=FALSE)
 #' AFM.lineProfile(afmd, 100, 500, 900, 500) -> afmd2
 #' AFM.linePlot(afmd2)
+#' 
+#' AFM.lineProfile(afmd, 1, 1, 128, 128, unitPixels=TRUE) -> afmd2
+#' AFM.linePlot(afmd2)
 #'
 #' @seealso \code{\link{AFM.getLine}}, \code{\link{AFM.linePlot}}
-#'
+#' @importFrom raster rasterFromXYZ click
+#' @importFrom sp plot
 #' @export
-AFM.lineProfile <- function(obj,x1,y1,x2,y2, verbose=FALSE) {
+AFM.lineProfile <- function(obj,x1=NA,y1=NA,x2=NA,y2=NA,
+                            unitPixels = FALSE, verbose=FALSE) {
   AFMcopy <- obj
   AFMcopy@history <- paste(AFMcopy@history,
                            "AFM.lineProfile(",x1,",",y1,",",x2,",",y2,");")
   d = AFM.raster(AFMcopy)
-  range.x = max(d$x) - min(d$x)
-  range.y = max(d$y) - min(d$y)
-  if(x1 >= range.x) { warning("x1: Out of range"); x1=0.99*range.x }
-  if(y1 >= range.y) { warning("y1: Out of range"); y1=0.99*range.y }
-  if(x2 >= range.x) { warning("x2: Out of range"); x2=0.99*range.x}
-  if(y2 >= range.y) { warning("y2: Out of range"); y2=0.99*range.y}
+  
   width.x = AFMcopy@x.pixels
   width.y = AFMcopy@y.pixels
-  x1.pixel = round(x1/range.x*width.x)
-  y1.pixel = round(y1/range.y*width.y)
-  x2.pixel = round(x2/range.x*width.x)
-  y2.pixel = round(y2/range.y*width.y)
+  
+  # if no coordinates are provided, use graphical interface
+  # to prompt for coordinates
+  if (is.na(x1) | is.na(x2) | is.na(y1) | is.na(y2)) {
+    dfr = raster::rasterFromXYZ(d)
+    sp::plot(dfr)
+    raster::click(dfr, n=2, xy=TRUE, show=FALSE) -> xy
+    x1 = xy$x[1]
+    x2 = xy$x[2]
+    y1 = xy$y[1]
+    y2 = xy$y[2]
+    # units are [nm] in this case
+    unitPixels = FALSE
+  }
+  
+  if (!unitPixels) {
+    range.x = max(d$x) - min(d$x)
+    range.y = max(d$y) - min(d$y)
+    if(x1 >= range.x) { warning("x1: Out of range"); x1=0.99*range.x }
+    if(y1 >= range.y) { warning("y1: Out of range"); y1=0.99*range.y }
+    if(x2 >= range.x) { warning("x2: Out of range"); x2=0.99*range.x}
+    if(y2 >= range.y) { warning("y2: Out of range"); y2=0.99*range.y}
+
+    x1.pixel = round(x1/range.x*width.x)
+    y1.pixel = round(y1/range.y*width.y)
+    x2.pixel = round(x2/range.x*width.x)
+    y2.pixel = round(y2/range.y*width.y)
+  } else {
+    if ((x1<0) | (x2<0) | (y1<0) | (y2<0)) { warning("Must have positive coordinates.") }
+    if (x1 > width.x) { warning("x1: Out of range"); x1 = width.x }
+    if (x2 > width.x) { warning("x2: Out of range"); x2 = width.x }
+    if (y1 > width.y) { warning("y1: Out of range"); y1 = width.y }
+    if (y2 > width.y) { warning("y2: Out of range"); y2 = width.y }
+    x1.pixel = x1
+    y1.pixel = y1
+    x2.pixel = x2
+    y2.pixel = y2
+  }
+  
   if (verbose) print(paste("Pixels: (",x1.pixel,",",y1.pixel,") - (",x2.pixel,",",y2.pixel,")"))
 
   Dx = abs(x2.pixel - x1.pixel)